fix: keep compatibility wiring within repository gates
